move joystick impl to src file and update how scaling operator is defined in vector
